+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add instructions for `PersistentNameID2Attribute` class
 - Add `RequiredAttributes` class
+- Include email address of the IdP in OpaqueSmartID error message
 
 ## [v1.0.0] - 2019-09-10